+++ apps/test-app/src/frontend/appui/frontstages/WidgetApiFrontstage.tsx
@@ -15,11 +15,14 @@ import {
   StageUsage,
   StandardContentLayouts,
   StandardLayout,
+  ToolbarComposer,
   ToolbarItemUtilities,
   ToolbarOrientation,
   ToolbarUsage,
+  ToolWidgetComposer,
   UiFramework,
   UiItemsProvider,
+  useActiveToolId,
   useConditionalValue,
   Widget,
   WidgetAction,
@@ -70,12 +73,6 @@ export function createWidgetApiFrontstage(): Frontstage {
         },
       ],
     },
-    cornerButton: (
-      <BackstageAppButton
-        key="appui-test-providers-WidgetApi-backstage"
-        icon="icon-bentley-systems"
-      />
-    ),
     defaultTool: MeasureDistanceTool.toolId,
     usage: StageUsage.General,
     topPanelProps: {
@@ -101,6 +98,10 @@ export function createWidgetApiFrontstage(): Frontstage {
 
   return {
     ...config,
+    contentManipulation: {
+      id: `${createWidgetApiFrontstage.stageId}-contentManipulationTools`,
+      content: <ContentManipulationWidget />,
+    },
     toolSettings,
     layout: (
       <StandardLayout
@@ -120,6 +121,52 @@ export function createWidgetApiFrontstage(): Frontstage {
   };
 }
 createWidgetApiFrontstage.stageId = "widget-api";
+
+function useActiveItemIds() {
+  const activeToolId = useActiveToolId();
+  const [showCustomViewOverlay, setShowCustomViewOverlay] = React.useState(
+    () => store.state.showCustomViewOverlay
+  );
+  React.useEffect(() => {
+    return store.onChanged.addListener(() => {
+      setShowCustomViewOverlay(store.state.showCustomViewOverlay);
+    });
+  }, []);
+  return React.useMemo(
+    () => [
+      ...(activeToolId ? [activeToolId] : []),
+      ...(showCustomViewOverlay
+        ? [createToggleCustomOverlayToolbarItem.id]
+        : []),
+    ],
+    [activeToolId, showCustomViewOverlay]
+  );
+}
+
+function ContentManipulationWidget() {
+  const activeItemIds = useActiveItemIds();
+  return (
+    <ToolWidgetComposer
+      cornerItem={<BackstageAppButton icon="icon-bentley-systems" />}
+      horizontalToolbar={
+        <ToolbarComposer
+          items={[]}
+          usage={ToolbarUsage.ContentManipulation}
+          orientation={ToolbarOrientation.Horizontal}
+          activeItemIds={activeItemIds}
+        />
+      }
+      verticalToolbar={
+        <ToolbarComposer
+          items={[]}
+          usage={ToolbarUsage.ContentManipulation}
+          orientation={ToolbarOrientation.Vertical}
+          activeItemIds={activeItemIds}
+        />
+      }
+    />
+  );
+}
 
 function RestoreLayoutWidgetAction() {
   return (
@@ -467,9 +514,10 @@ function createToggleCustomOverlayToolbarItem() {
     );
   };
   return ToolbarItemUtilities.createActionItem({
-    id: "testHideShowItems",
+    id: createToggleCustomOverlayToolbarItem.id,
     icon: <CustomOverlayIcon />,
     label,
     execute,
   });
 }
+createToggleCustomOverlayToolbarItem.id = "testHideShowItems";